+++ projects/ids-enterprise-ng/src/lib/lookup/soho-lookup.component.ts
@@ -343,6 +343,9 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
 
   @Input() toolbar: any;
 
+  // Autocomplete template settings
+  @Input() autoCompleteSettings?: SohoLookupAutoComplete;
+
   /**
    * Available Soho Template events as Output (EventEmitters passing the event)
    * Should match the Soho event names for the component
@@ -434,6 +437,11 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
    */
   private lookup?: SohoLookupStatic | null;
 
+  /**
+   * Autocomplete
+   */
+  private autocomplete!: SohoAutoCompleteStatic | null;
+
   private settings: SohoLookupOptions = {};
 
   /** Initial dataset */
@@ -458,6 +466,7 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
   ngAfterViewInit() {
     this.ngZone.runOutsideAngular(() => {
       this.jQueryElement = jQuery(this.element.nativeElement);
+      const dataSet = this._dataset ? this._dataset : [];
 
       // The default options for the data grid, which will
       // be overriden by the options provided by the caller
@@ -465,7 +474,7 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
       const datagridConfig: SohoDataGridOptions = {
         cellNavigation: false,
         columns: this.columns,
-        dataset: this._dataset ? this._dataset : [],
+        dataset: dataSet,
         selectable: this.isMultiselect() ? 'multiple' : 'single',
         toolbar: Object.assign({
           actions: true,
@@ -482,6 +491,27 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
 
       this.settings.options = Object.assign(datagridConfig, this.options);
       this.jQueryElement.lookup(this.settings);
+
+      // setup autocomplete
+      if (this.autoCompleteSettings) {
+        const acDataset = dataSet.map((data: { [key: string]: any }) => {
+          return this.autoCompleteSettings ? {
+            id: data[this.autoCompleteSettings.id],
+            value: data[this.autoCompleteSettings.value],
+            label: data[this.autoCompleteSettings.label]
+          } : data;
+        });
+
+        const acSettings: SohoAutoCompleteOptions = { source: acDataset };
+
+        if (this.autoCompleteSettings.template) {
+          acSettings.template = this.autoCompleteSettings.template;
+        }
+
+        this.jQueryElement.autocomplete(acSettings);
+
+        this.autocomplete = this.jQueryElement.data('autocomplete');
+      }
 
       /**
        * Bind to jQueryElement's events
@@ -604,9 +634,15 @@ export class SohoLookupComponent extends BaseControlValueAccessor<any> implement
       if (this.jQueryElement) {
         this.jQueryElement.off();
       }
+
       if (this.lookup) {
         this.lookup.destroy();
         this.lookup = null;
+      }
+
+      if (this.autocomplete) {
+        this.autocomplete.destroy();
+        this.autocomplete = null;
       }
     });
   }